remove internal/pkg/ dir
