+++ internal/command/httpserver.go
@@ -8,7 +8,7 @@ import (
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
 
-	"github.com/elastic/stream/internal/pkg/httpserver"
+	"github.com/elastic/stream/internal/httpserver"
 )
 
 type httpServerRunner struct {
