+++ internal/output/azureblobstorage/azure_blob_storage_test.go
@@ -2,7 +2,7 @@
 // Elasticsearch B.V. licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information.
 
-package gcs
+package azureblobstorage
 
 import (
 	"context"
@@ -10,27 +10,25 @@ import (
 	"fmt"
 	"io"
 	"log"
-	"net"
 	"net/http"
 	"os"
 	"testing"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
 	"github.com/ory/dockertest/v3"
 	"github.com/ory/dockertest/v3/docker"
 	"github.com/stretchr/testify/require"
 	"gotest.tools/assert"
 
-	"github.com/elastic/stream/internal/pkg/output"
+	"github.com/elastic/stream/internal/output"
 )
 
 const (
-	emulatorHost = "localhost"
-	emulatorPort = "4443"
-	bucket       = "testbucket"
-	objectname   = "testobject"
+	emulatorHost = "127.0.0.1"
+	emulatorPort = "10000"
+	container    = "testcontainer"
+	blob         = "testblob"
 )
-
-var emulatorHostAndPort = fmt.Sprintf("http://%s", net.JoinHostPort(emulatorHost, emulatorPort))
 
 func TestMain(m *testing.M) {
 	pool, err := dockertest.NewPool("")
@@ -39,9 +37,8 @@ func TestMain(m *testing.M) {
 	}
 
 	resource, err := pool.RunWithOptions(&dockertest.RunOptions{
-		Repository: "fsouza/fake-gcs-server",
+		Repository: "mcr.microsoft.com/azure-storage/azurite",
 		Tag:        "latest",
-		Cmd:        []string{"-host=0.0.0.0", "-public-host=localhost", fmt.Sprintf("-port=%s", emulatorPort), "-scheme=http"},
 		PortBindings: map[docker.Port][]docker.PortBinding{
 			emulatorPort: {{HostIP: emulatorHost, HostPort: emulatorPort}},
 		},
@@ -60,25 +57,21 @@ func TestMain(m *testing.M) {
 	if err := pool.Retry(func() error {
 		// Disable HTTP keep-alives to ensure no extra goroutines hang around.
 		httpClient := http.Client{Transport: &http.Transport{DisableKeepAlives: true}}
-
 		// Sanity check the emulator.
-		resp, err := httpClient.Get(fmt.Sprintf("http://%s:%s/storage/v1/b", emulatorHost, emulatorPort))
+		resp, err := httpClient.Get(fmt.Sprintf("http://%s:%s/devstoreaccount1", emulatorHost, emulatorPort))
 		if err != nil {
 			return err
 		}
 		defer resp.Body.Close()
 
-		_, err = io.ReadAll(resp.Body)
-		if err != nil {
-			return err
-		}
-		if resp.StatusCode != http.StatusOK {
+		if resp.StatusCode != http.StatusBadRequest {
 			return fmt.Errorf("unexpected status code: %v", resp.StatusCode)
 		}
-		return err
+
+		return nil
 	}); err != nil {
 		_ = pool.Purge(resource)
-		log.Fatalf("Could not connect to the gcs instance: %s", err)
+		log.Fatalf("Could not connect to the Azure Blob Storage instance: %s", err)
 	}
 
 	code := m.Run()
@@ -88,12 +81,13 @@ func TestMain(m *testing.M) {
 	os.Exit(code)
 }
 
-func TestGcs(t *testing.T) {
+func TestAzureBlobStorage(t *testing.T) {
 	out, err := New(&output.Options{
-		Addr: emulatorHostAndPort,
-		GcsOptions: output.GcsOptions{
-			Bucket: bucket,
-			Object: objectname,
+		Addr: emulatorHost,
+		AzureBlobStorageOptions: output.AzureBlobStorageOptions{
+			Container: container,
+			Blob:      blob,
+			Port:      emulatorPort,
 		},
 	})
 	require.NoError(t, err)
@@ -111,21 +105,20 @@ func TestGcs(t *testing.T) {
 	require.NoError(t, err)
 	assert.Equal(t, len(data), n)
 
-	// Need to close the Writer for the Object to be created in the Bucket.
-	out.Close()
+	connectionString := fmt.Sprintf("DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://%s:%s/devstoreaccount1;", emulatorHost, emulatorPort)
+	ctx, cancel := context.WithCancel(context.Background())
+	serviceClient, _ := azblob.NewClientFromConnectionString(connectionString, nil)
 
-	gcsClient, ctx, cancel, err := NewClient(emulatorHostAndPort)
+	blobDownloadResponse, err := serviceClient.DownloadStream(ctx, container, blob, nil)
 	require.NoError(t, err)
-	t.Cleanup(func() { _ = gcsClient.Close() })
+
+	reader := blobDownloadResponse.Body
+	downloadData, err := io.ReadAll(reader)
+	require.NoError(t, err)
+	assert.Equal(t, string(data), string(downloadData))
+
+	err = reader.Close()
+	require.NoError(t, err)
+
 	t.Cleanup(cancel)
-
-	o := gcsClient.Bucket(bucket).Object(objectname)
-	r, err := o.NewReader(ctx)
-	require.NoError(t, err)
-
-	body, err := io.ReadAll(r)
-	require.NoError(t, err)
-	defer r.Close()
-
-	assert.Equal(t, string(data), string(body))
 }
