+++ internal/output/factory.go
