+++ internal/output/tcp/tcp.go
@@ -9,7 +9,7 @@ import (
 	"net"
 	"time"
 
-	"github.com/elastic/stream/internal/pkg/output"
+	"github.com/elastic/stream/internal/output"
 )
 
 func init() {
