+++ internal/output/tls/tls.go
@@ -10,7 +10,7 @@ import (
 	"net"
 	"time"
 
-	"github.com/elastic/stream/internal/pkg/output"
+	"github.com/elastic/stream/internal/output"
 )
 
 func init() {
