+++ internal/output/webhook/webhook.go
@@ -14,7 +14,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/elastic/stream/internal/pkg/output"
+	"github.com/elastic/stream/internal/output"
 )
 
 func init() {
